Removes Boron-10 and Antimony. Adds an empty hall to build_hall.sh
Boron-10 and Antimony are not in the list of nuclei that are included in FNAL's "small" GENIE cross-section splines. This commit removes them to avoid the hassle. Antimony is not used in the current geometry. Boron is now just Boron-11, the more common isotope. There is almost no Boron in the detectors anyway, so this change will never matter.

build_hall.sh now also builds an empty hall (nd_hall_no_dets.gdml) along with the hall with detectors (nd_hall_with_dets.gdml).